Move 'install --self' to a new 'selfinstall' command
The selfinstall command now houses all dnvm installation/copying
behavior, while install is isolated to handling SDKs.
