+++ src/dnvm/BannedSymbols.txt
@@ -1,0 +1,1 @@
+M:Zio.UPath.UPath(System.String);Don't use implicit conversion
