+++ src/dnvm/List.cs
@@ -8,7 +8,7 @@ public static class ListCommand
 {
     /// <summary>
     /// Prints a list of installed SDK versions and their locations.
-    public static Task<int> Run(Logger logger, DnvmHome home)
+    public static Task<int> Run(Logger logger, DnvmFs home)
     {
         Manifest manifest;
         try
