+++ src/dnvm/dnvm.csproj
@@ -15,12 +15,20 @@
     <PackageReference Include="GitVersion.MSBuild" Version="5.10.3">
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
+    <PackageReference Include="Microsoft.CodeAnalysis.BannedApiAnalyzers" Version="3.3.4">
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+      <PrivateAssets>all</PrivateAssets>
+    </PackageReference>
     <PackageReference Include="Semver" Version="2.2.0" />
 
     <PackageReference Include="Serde" Version="0.4.6" />
     <PackageReference Include="StaticCs" Version="0.2.0" />
     <PackageReference Include="StaticCs.Collections" Version="1.0.1" />
     <PackageReference Include="zio" Version="0.16.2" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <AdditionalFiles Include="BannedSymbols.txt" />
   </ItemGroup>
 
 </Project>
